adds OwnedMeasurement::set_measurement_name, a workaround for setting dynamic measurement names
with measure!, there is no way to use the value of a name, the name itself ends up as the `key` field value.
this is part of the reason it's nice to use measure!. when you don't know the measurement name at compile time,
there is no way to use a dynamic value using measure!. this provides a way to modify the measurement name
after constructing the OwnedMeasurement (you can use measure!(@make_meas ..) for that).